added oauth Google authentication

+++ controllers/api/v1/post_api.js
@@ -20,12 +20,18 @@ module.exports.destroy = async (req, res) => {
   try {
     let post = await Post.findById(req.params.id);
 
-    post.remove();
-    await Comment.deleteMany({ post: req.params.id });
+    if (post.user == req.user.id) {
+      post.remove();
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: "Post and Associated comments deleted",
-    });
+      return res.json(200, {
+        message: "Post and Associated comments deleted",
+      });
+    } else {
+      return res.json(401, {
+        message: "you can not delete this post",
+      });
+    }
   } catch (error) {
     return res.json(500, {
       message: "Internal server error",
